Remove xbuild support
xbuild is deprecated and msbuild support is solid, there's
no reason to support xbuild any longer.

Getting rid of this baggage allows us to depend on things that
only work in MSBuild.
