+++ main/src/core/MonoDevelop.Core/MonoDevelop.Projects.MSBuild/SdkResolution.cs
@@ -104,7 +104,7 @@ namespace MonoDevelop.Projects.MSBuild
 			// Also add the default resolver.
 
 			var resolvers = new List<SdkResolver> { new MonoDevelop.Projects.MSBuild.Resolver (MSBuildProjectService.FindRegisteredSdks), new DefaultSdkResolver { TargetRuntime = runtime } };
-			MSBuildProjectService.GetNewestInstalledToolsVersion (runtime, true, out var binDir);
+			var binDir = MSBuildProjectService.GetMSBuildBinPath (runtime);
 			var potentialResolvers = FindPotentialSdkResolvers (Path.Combine (binDir, "SdkResolvers"));
 
 			if (potentialResolvers.Count == 0) return resolvers;
