+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Gui.OptionPanels/BuildPanel.cs
@@ -63,7 +63,6 @@ namespace MonoDevelop.Ide.Gui.OptionPanels
 			buildBeforeRunCheckBox.Active = IdeApp.Preferences.BuildBeforeExecuting;
 			verbosityCombo.Active = (int)IdeApp.Preferences.MSBuildVerbosity.Value;
 			buildBeforeTestCheckBox.Active = IdeApp.Preferences.BuildBeforeRunningTests;
-			buildWithMSBuildCheckBox.Active = Runtime.Preferences.BuildWithMSBuild;
 			skipBuildingUnmodifiedProjectsCheckbox.Active = Runtime.Preferences.SkipBuildingUnmodifiedProjects;
 			parallelBuildCheckbox.Active = MonoDevelop.Core.Runtime.Preferences.ParallelBuild.Value;
 
@@ -78,8 +77,6 @@ namespace MonoDevelop.Ide.Gui.OptionPanels
 			                                                          GettextCatalog.GetString ("Check to run the solution even if the build had warnings"));
 			buildBeforeTestCheckBox.SetCommonAccessibilityAttributes ("BuildPanel.buildBeforeTest", "",
 			                                                          GettextCatalog.GetString ("Check to build the solution before running tests"));
-			buildWithMSBuildCheckBox.SetCommonAccessibilityAttributes ("BuildPanel.buildWithMSBuild", "",
-			                                                           GettextCatalog.GetString ("Check to use MSBuild to build the solution"));
 			skipBuildingUnmodifiedProjectsCheckbox.SetCommonAccessibilityAttributes ("BuildPanel.skipUnmodifiedProject", "",
 																					 GettextCatalog.GetString ("Check to skip building unmodified projects"));
 			parallelBuildCheckbox.SetCommonAccessibilityAttributes ("BuildPanel.parallelBuild", "",
@@ -102,7 +99,6 @@ namespace MonoDevelop.Ide.Gui.OptionPanels
 			IdeApp.Preferences.BuildBeforeExecuting.Value = buildBeforeRunCheckBox.Active;
 			IdeApp.Preferences.MSBuildVerbosity.Value = (MSBuildVerbosity) verbosityCombo.Active;
 			IdeApp.Preferences.BuildBeforeRunningTests.Value = buildBeforeTestCheckBox.Active;
-			Runtime.Preferences.BuildWithMSBuild.Value = buildWithMSBuildCheckBox.Active;
 			Runtime.Preferences.SkipBuildingUnmodifiedProjects.Value = skipBuildingUnmodifiedProjectsCheckbox.Active;
 			MonoDevelop.Core.Runtime.Preferences.ParallelBuild.Value = parallelBuildCheckbox.Active;
 			if (saveChangesRadioButton.Active)
