+++ main/src/core/MonoDevelop.Projects.Formats.MSBuild/AssemblyInfo.cs
@@ -37,6 +37,8 @@ using System.Runtime.CompilerServices;
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
+[assembly: System.Reflection.AssemblyTitle ("MonoDevelop MSBuild Builder")]
+
 // The assembly version has the format "{Major}.{Minor}.{Build}.{Revision}".
 // The form "{Major}.{Minor}.*" will automatically update the build and revision,
 // and "{Major}.{Minor}.{Build}.*" will update just the revision.
@@ -48,3 +50,8 @@ using System.Runtime.CompilerServices;
 
 //[assembly: AssemblyDelaySign(false)]
 //[assembly: AssemblyKeyFile("")]
+
+class MSBuildConsts
+{
+	public const string Version = "15.0";
+}
