--- main/src/core/MonoDevelop.Projects.Formats.MSBuild/AssemblyInfo.v12.0.cs
@@ -1,9 +1,0 @@
-[assembly: System.Reflection.AssemblyTitle ("MonoDevelop MSBuild 12.0 Builder")]
-
-namespace MonoDevelop.Projects.MSBuild
-{
-	static class MSBuildConsts
-	{
-		public const string Version = "12.0";
-	}
-}
