+++ main/src/core/MonoDevelop.Projects.Formats.MSBuild/Makefile.am
@@ -1,4 +1,1 @@
 include $(top_srcdir)/xbuild.include
-
-SUBDIRS = \
-	MonoDevelop.Projects.MSBuild.Shared
