+++ main/src/core/MonoDevelop.Projects.Formats.MSBuild/MonoDevelop.MSBuildBuilder.csproj
@@ -8,16 +8,15 @@
     <SchemaVersion>2.0</SchemaVersion>
     <ProjectGuid>{7F5B649A-3572-4713-83FD-C28A6AA70445}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>MonoDevelop.Projects.Formats.MSBuild</RootNamespace>
-    <AssemblyName>MonoDevelop.Projects.Formats.MSBuild</AssemblyName>
+    <RootNamespace>MonoDevelop.Projects.MSBuild</RootNamespace>
+    <AssemblyName>MonoDevelop.MSBuildBuilder</AssemblyName>
     <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
-    <BaseIntermediateOutputPath>obj\dotnet.15.0\</BaseIntermediateOutputPath>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|x86' ">
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
-    <OutputPath>..\..\..\build\bin\MSBuild\dotnet.15.0</OutputPath>
+    <OutputPath>..\..\..\build\bin</OutputPath>
     <DefineConstants>DEBUG;</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
@@ -28,7 +27,7 @@
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
-    <OutputPath>..\..\..\build\bin\MSBuild\dotnet.15.0</OutputPath>
+    <OutputPath>..\..\..\build\bin</OutputPath>
     <DefineConstants>DEBUG;</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
@@ -38,7 +37,7 @@
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|x86' ">
     <DebugType></DebugType>
     <Optimize>true</Optimize>
-    <OutputPath>..\..\..\build\bin\MSBuild\dotnet.15.0</OutputPath>
+    <OutputPath>..\..\..\build\bin</OutputPath>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <ExternalConsole>true</ExternalConsole>
@@ -46,7 +45,7 @@
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'ReleaseWin32|x86' ">
     <Optimize>true</Optimize>
-    <OutputPath>..\..\..\build\bin\MSBuild\dotnet.15.0</OutputPath>
+    <OutputPath>..\..\..\build\bin</OutputPath>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <ExternalConsole>true</ExternalConsole>
@@ -87,28 +86,43 @@
     </Reference>
   </ItemGroup>
   <ItemGroup Condition=" '$(Configuration)' == 'Debug' OR '$(Configuration)' == 'Release' ">
-    <None Include="app.v15.0.config">
+    <None Include="app.config">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <Link>MonoDevelop.Projects.Formats.MSBuild.exe.config</Link>
+      <Link>MonoDevelop.MSBuildBuilder.exe.config</Link>
     </None>
   </ItemGroup>
   <ItemGroup Condition=" '$(Configuration)' == 'DebugWin32' OR '$(Configuration)' == 'ReleaseWin32' ">
-    <None Include="app.v15.0.windows.config">
+    <None Include="app.windows.config">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <Link>MonoDevelop.Projects.Formats.MSBuild.exe.config</Link>
+      <Link>MonoDevelop.MSBuildBuilder.exe.config</Link>
     </None>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="MonoDevelop.Projects.Formats.MSBuild\BuildEngine.v4.0.cs" />
-    <Compile Include="MonoDevelop.Projects.Formats.MSBuild\LocalLogger.cs" />
-    <Compile Include="MonoDevelop.Projects.Formats.MSBuild\ProjectBuilder.v4.0.cs" />
-    <Compile Include="AssemblyInfo.v15.0.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\BuildEngine.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\LocalLogger.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\ProjectBuilder.cs" />
     <Compile Include="AssemblyInfo.cs" />
-    <Compile Include="MonoDevelop.Projects.Formats.MSBuild\MSBuildLoggerAdapter.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\MSBuildLoggerAdapter.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\Main.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\BuildEngine.Shared.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\LoggerAdapter.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\RemoteBuildEngineMessages.cs" />
+    <Compile Include="..\MonoDevelop.Core\MonoDevelop.Core.Execution\BinaryMessage.cs">
+      <Link>MonoDevelop.Core.Execution\BinaryMessage.cs</Link>
+    </Compile>
+    <Compile Include="MonoDevelop.Projects.MSBuild\ILogWriter.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\MSBuildEvaluatedItem.cs" />
+    <Compile Include="..\MonoDevelop.Core\MonoDevelop.Projects.MSBuild\MSBuildEvent.cs">
+      <Link>MonoDevelop.Core.Execution\MSBuildEvent.cs</Link>
+    </Compile>
+    <Compile Include="MonoDevelop.Projects.MSBuild\MSBuildResult.cs" />
+    <Compile Include="MonoDevelop.Projects.MSBuild\MSBuildTargetResult.cs" />
+    <Compile Include="..\MonoDevelop.Core\MonoDevelop.Core.Execution\RemoteProcessServer.cs">
+      <Link>MonoDevelop.Core.Execution\RemoteProcessServer.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
-  <Import Project="MonoDevelop.Projects.MSBuild.Shared\MonoDevelop.Projects.MSBuild.Shared.projitems" Label="Shared" Condition="Exists('MonoDevelop.Projects.MSBuild.Shared\MonoDevelop.Projects.MSBuild.Shared.projitems')" />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
 </Project>
