--- main/src/core/MonoDevelop.Projects.Formats.MSBuild/MonoDevelop.Projects.MSBuild.Shared/Makefile.am
@@ -1,1 +1,0 @@
-include $(top_srcdir)/xbuild.include
