--- main/src/core/MonoDevelop.Projects.Formats.MSBuild/MonoDevelop.Projects.MSBuild.Shared/MonoDevelop.Projects.MSBuild.Shared.projitems
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>{DB95FF74-275D-4488-84E9-1060527988BF}</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>MonoDevelop.Projects.MSBuild.Shared</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)..\..\MonoDevelop.Core\MonoDevelop.Core.Execution\BinaryMessage.cs">
-      <Link>BinaryMessage.cs</Link>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)BuildEngine.Shared.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)RemoteBuildEngineMessages.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ProjectBuilder.Shared.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)..\..\MonoDevelop.Core\MonoDevelop.Core.Execution\RemoteProcessServer.cs">
-      <Link>RemoteProcessServer.cs</Link>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)ILogWriter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)MSBuildEvaluatedItem.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)MSBuildResult.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)MSBuildTargetResult.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Main.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)LocalLogger.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)..\..\MonoDevelop.Core\MonoDevelop.Projects.MSBuild\MSBuildEvent.cs">
-      <Link>MSBuildEvent.cs</Link>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)LoggerAdapter.cs" />
-  </ItemGroup>
-</Project>
