+++ main/src/core/MonoDevelop.Projects.Formats.MSBuild/MonoDevelop.Projects.MSBuild/MSBuildEvaluatedItem.cs
