+++ main/src/core/MonoDevelop.Projects.Formats.MSBuild/app.config
