--- main/src/core/MonoDevelop.Projects.Formats.MSBuild/app.v4.0.config
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-	<runtime>
-		<AppContextSwitchOverrides value="Switch.System.IO.UseLegacyPathHandling=false" />
-		<generatePublisherEvidence enabled="false" />
-		<assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-			<dependentAssembly>
-				<assemblyIdentity name="Microsoft.Build.Framework" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
-				<bindingRedirect oldVersion="0.0.0.0-100.0.0.0" newVersion="4.0.0.0" />
-			</dependentAssembly>
-			<dependentAssembly>
-				<assemblyIdentity name="Microsoft.Build.Engine" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
-				<bindingRedirect oldVersion="0.0.0.0-100.0.0.0" newVersion="4.0.0.0" />
-			</dependentAssembly>
-		</assemblyBinding>
-	</runtime>
-</configuration>
