+++ main/src/core/MonoDevelop.Projects.Formats.MSBuild/app.windows.config
